move all files needed for test in same folder
